Add Function separation
