+++ packages/hello-gsm/src/components/CalculatorPage/function/Rounds.ts
@@ -1,0 +1,5 @@
+const Rounds = (num: number) => {
+  return Number(num.toFixed(4));
+}; // 소수점 넷째 자리에서 반올림 후 형변환
+
+export default Rounds;
