+++ packages/hello-gsm/src/components/CalculatorPage/function/index.ts
@@ -1,0 +1,4 @@
+export { default as Volunteer } from './Volunteer';
+export { default as Calculate } from './Calculate';
+export { default as Rounds } from './Rounds';
+export { default as Attendance } from './Attendance';
