+++ packages/hello-gsm/src/components/CalculatorPage/index.tsx
@@ -4,6 +4,7 @@ import * as S from './style';
 import * as I from '../../Assets/svg';
 import { useForm } from 'react-hook-form';
 import { useState } from 'react';
+import { Calculate, Volunteer, Rounds, Attendance } from './function';
 
 interface ScoreForm {
   score2_1: number[];
@@ -16,90 +17,32 @@ interface ScoreForm {
 }
 
 const CalculatorPage: NextPage = () => {
-  const { register, handleSubmit, watch, control } = useForm<ScoreForm>();
-
-  const Rounds = (num: number) => {
-    return Number(num.toFixed(4));
-  }; // 소수점 넷째 자리에서 반올림 후 형변환
-
-  // 출석점수 환산
-  const Attendance = (absent: Array<number>, attendance: Array<number>) => {
-    let result = 30;
-    absent.map((arr, i) => {
-      absent[i] = Number(arr);
-    }); // 배열을 number형변환
-    attendance.map((arr, i) => {
-      attendance[i] = Number(arr);
-    }); // 배열을 number형변환
-    const absentSum = absent.reduce((accumulator, curr) => accumulator + curr); // 배열 총합계 저장
-    if (absentSum >= 10) return 0;
-
-    const attendanceSum = Math.floor(
-      attendance.reduce((accumulator, curr) => accumulator + curr) / 3,
-    ); // 배열 총합계 저장, 소수점 버림
-
-    result = result - 3 * (absentSum + attendanceSum);
-    if (result < 0) return 0; // 결과가 음수이면 0 반환
-    return result;
-  };
-
-  // 봉사점수 환산
-  const Volunteer = (array: Array<number>) => {
-    let result = 0;
-    array.map((arr, i) => {
-      const score = array[i];
-      if (score >= 7) result += 10;
-      else if (score >= 6) result += 8;
-      else if (score >= 5) result += 6;
-      else if (score >= 4) result += 4;
-      else result += 2;
-    });
-    return result;
-  };
-
-  // 성적 계산
-  const Calculate = (array: Array<number>, grade: number) => {
-    array.map((arr, i) => {
-      array[i] = Number(arr);
-    }); // 배열을 number형변환
-
-    const sum = array.reduce((accumulator, curr) => accumulator + curr); // 배열 총합계 저장
-    let result = 0; // 결과
-    switch (grade) {
-      case 2:
-        result = 54 * (sum / (array.length * 5)); // 2학년 공식 적용
-        break;
-      case 3:
-        result = 72 * (sum / (array.length * 5)); // 3학년 공식 적용
-        break;
-      case 4:
-        result = 60 * (sum / (array.length * 5)); // 예체능 공식 적용
-        break;
-    }
-    return Rounds(result);
-  };
+  const { register, handleSubmit } = useForm<ScoreForm>();
 
   const onValid = (validForm: ScoreForm) => {
-    const score2_1 = Calculate(validForm.score2_1, 2); // 2학년 1학기
-    const score2_2 = Calculate(validForm.score2_2, 2); // 2학년 2학기
-    const score3_1 = Calculate(validForm.score3_1, 3); // 3학년 1학기
+    const score2_1: number = Calculate(validForm.score2_1, 2); // 2학년 1학기
+    const score2_2: number = Calculate(validForm.score2_2, 2); // 2학년 2학기
+    const score3_1: number = Calculate(validForm.score3_1, 3); // 3학년 1학기
 
-    const generalCurriculumScoreSubtotal = Rounds(
+    const generalCurriculumScoreSubtotal: number = Rounds(
       score2_1 + score2_2 + score3_1,
-    ); // 소계
-    const artSportsScore = Calculate(validForm.artSportsScore, 4); // 예체능
-    const attendanceScore = Attendance(
+    ); // 교과성적 소계
+    const artSportsScore: number = Calculate(validForm.artSportsScore, 4); // 예체능
+    const attendanceScore: number = Attendance(
       validForm.absentScore,
       validForm.attendanceScore,
     ); // 출석점수
-    const volunteerScore = Volunteer(validForm.volunteerScore); // 봉사점수
-    const curriculumScoreSubtotal = Rounds(
+
+    const volunteerScore: number = Volunteer(validForm.volunteerScore); // 봉사점수
+    const curriculumScoreSubtotal: number = Rounds(
       generalCurriculumScoreSubtotal + artSportsScore,
     );
-    const nonCurriculumScoreSubtotal = Rounds(attendanceScore + volunteerScore); //비교과 성적 소계
-    const scoreTotal = Rounds(
+    const nonCurriculumScoreSubtotal: number = Rounds(
+      attendanceScore + volunteerScore,
+    ); //비교과 성적 소계
+    const scoreTotal: number = Rounds(
       curriculumScoreSubtotal + nonCurriculumScoreSubtotal,
-    );
+    ); // 총합
   };
 
   const lines = ['일반교과', '체육•예술 교과', '비교과'];
